🚚 rename `status` file endpoint to `ping`
